+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Enhancements:
 Bugfix:
   - Edit slider min value #420
   - Fix extent error when user clicks in the viewer to create a box by clicking outside of the image #425
+  - Fix extent error when user enlarges the box outside the image #425
+  - Deprecates `GetImageWorldExtent` #425
 
 ## v24.0.1